+++ src/contraction_sequences.jl
@@ -1,9 +1,10 @@
 using Graphs: vertices
-using ITensors: ITensor, contract
+using ITensors: ITensor, contract, inds, dim
 using ITensors.ContractionSequenceOptimization: deepmap
 using ITensors.NDTensors: Algorithm, @Algorithm_str
 using NamedGraphs.Keys: Key
 using NamedGraphs.OrdinalIndexing: th
+using TensorOperations: optimaltree
 
 function contraction_sequence(tn::Vector{ITensor}; alg="optimal", kwargs...)
   return contraction_sequence(Algorithm(alg), tn; kwargs...)
@@ -19,4 +20,11 @@ end
 
 function contraction_sequence(::Algorithm"optimal", tn::Vector{ITensor})
   return optimal_contraction_sequence(tn)
+end
+
+function ITensors.optimal_contraction_sequence(tensors::Vector{<:ITensor})
+  network = collect.(inds.(tensors))
+  inds_to_dims = Dict(i => dim(i) for i in unique(reduce(vcat, network)))
+  seq, _ = optimaltree(network, inds_to_dims)
+  return seq
 end
